Set environment variables in %environment
Remove the `/openrobots_setup.bash` and set the environment variables in
the `%environment`section instead.  This way they are set automatically
and the user does not need to manually source the file.

Note that the `/setup.bash` is still needed as it defines functions and
aliases which do not work in `%environment`.