Add "challenge image" containing our software
Add a `challenge_image.def` that extends an locally exiting blmc_ei.sif
by adding a catkin workspace with the BLMC_EI project built.
Further it defines two "apps" `build` and `run` which are going to be
used by the submission system.

In blmc_ei.def change the `catbulid` alias to be a function, so it can
be used in non-interactive scripts.