Use a template for blmc_ei.def
To be more flexible regarding base image and build options, use a
template for `blmc_ei.def` with some variables.  Handle this in the
Makefile, so building the images stays simple.

For the "normal" version of blmc_ei.sif (now called blmc_ei_user.sif)
run

    make blmc_ei_user.sif

For the "robot" version (including pylon and using real-time build):

    make blmc_ei_robots.sif

The makefile will automatically create a corresponding def file for each
target based on the template.