Clone workspace in Makefile instead of def %setup
Instead of cloning in the `%setup` section, clone the workspace to the
current working directory via the `workspace` target of the Makefile and
copy it from there to the image.

This makes it easier to create an image with a custom workspace and
avoids cloning the repositories twice when building both the "user" and
the "robot" image.