[base_pylon] Verify pylon package using sha256
We are downloading the pylon SDK package from an unofficial server
(since the official download on Basler's website is hidden behind a form
and thus not usable in a script) so there is a potential risk that the
package was modified (with malicious intend or maybe just by accident).
To make sure the downloaded package is save (i.e. the same as when
downloading from Basler), check its sha256 checksum.

The checksum used here was acquired by downloading the package from the
Basler website
(https://www.baslerweb.com/en/downloads/software-downloads/#type=pylonsoftware;language=all;version=5.2.0;os=linuxx8664bit).