Use persistent cache-directories
Bind directories for cache of apt-get, pip and wget so downloaded files
can be reused between builds.
This saves almost 1GB of downloads on the second build, thus saving both
time and energy (with my not-super-fast internet connection at home,
this reduces build time of the base image from ~9min to ~6min).