fix dependency on build/cache
The timestamp of `build/cache` changes if files are added to it.  Thus,
any build that uses the cache may "update" that target and thus trigger
a rebuild of the base image (which depends on `build/cache`).
This is unnecessary and actually results in the automated build of all
images.

To avoid that, specify the dependency on `build/cache` after a `|`.
This tells make to only consider the dependency for the build order but
ignore its timestamp (see https://stackoverflow.com/a/51082926).